refactor: Remove duplicate type definition for Icon
